Moved sidebar-menu into separate file
