+++ src/components/Sidebar.vue
@@ -28,85 +28,18 @@
       <!-- /.search form -->
 
       <!-- Sidebar Menu -->
-      <ul class="sidebar-menu">
-        <li class="header">TOOLS</li>
-        <li class="active pageLink" v-on:click="toggleMenu">
-          <router-link to="/"><i class="fa fa-desktop"></i>
-            <span class="page">Dashboard</span>
-          </router-link>
-        </li>
-        <li class="pageLink" v-on:click="toggleMenu">
-          <router-link to="/tables"><i class="fa fa-table"></i>
-            <span class="page">Tables</span>
-          </router-link>
-        </li>
-
-        <li class="header">ME</li>
-        <li class="pageLink" v-on:click="toggleMenu">
-          <router-link to="/tasks">
-            <i class="fa fa-tasks"></i>
-            <span class="page">Tasks</span>
-          </router-link>
-        </li>
-        <li class="pageLink" v-on:click="toggleMenu">
-          <router-link to="/setting">
-            <i class="fa fa-cog"></i>
-            <span class="page">Settings</span>
-          </router-link>
-        </li>
-
-        <li class="header">LOGS</li>
-        <li class="pageLink" v-on:click="toggleMenu">
-          <router-link to="/access"><i class="fa fa-book"></i>
-            <span class="page">Access</span>
-          </router-link>
-        </li>
-        <li class="pageLink" v-on:click="toggleMenu">
-          <router-link to="/server"><i class="fa fa-hdd-o"></i>
-            <span class="page">Server</span>
-          </router-link>
-        </li>
-        <li class="pageLink" v-on:click="toggleMenu">
-          <router-link to="/repos"><i class="fa fa-heart"></i>
-            <span class="page">Repos</span>
-            <small class="label pull-right bg-green">AJAX</small>
-          </router-link>
-        </li>
-
-        <li class="header">PAGES</li>
-        <li class="pageLink" v-on:click="toggleMenu">
-          <router-link to="/login">
-            <i class="fa fa-circle-o text-yellow"></i>
-            <span class="page"> Login</span>
-          </router-link>
-        </li>
-        <li class="pageLink" v-on:click="toggleMenu">
-          <router-link to="/404"><i class="fa fa-circle-o text-red"></i>
-            <span class="page"> 404</span>
-          </router-link>
-        </li>
-      </ul>
+      <sidebar-menu />
       <!-- /.sidebar-menu -->
     </section>
     <!-- /.sidebar -->
   </aside>
 </template>
 <script>
+import SidebarMenu from './SidebarMenu'
+
 export default {
   name: 'Sidebar',
   props: ['displayName'],
-  methods: {
-    toggleMenu (event) {
-      // remove active from li
-      var active = document.querySelector('li.pageLink.active')
-
-      if (active) {
-        active.classList.remove('active')
-      }
-      // window.$('li.pageLink.active').removeClass('active')
-      // Add it to the item that was clicked
-      event.toElement.parentElement.className = 'pageLink active'
-    }
-  }
+  components: { SidebarMenu }
 }
 </script>
