+++ src/main.js
@@ -53,8 +53,10 @@ new Vue({
 
 // Check local storage to handle refreshes
 if (window.localStorage) {
-  if (store.state.user !== window.localStorage.getItem('user')) {
-    store.commit('SET_USER', JSON.parse(window.localStorage.getItem('user')))
+  var localUserString = window.localStorage.getItem('user')
+
+  if (store.state.user !== localUserString) {
+    store.commit('SET_USER', JSON.parse(localUserString))
     store.commit('SET_TOKEN', window.localStorage.getItem('token'))
   }
 }
